refactor: move tapErr to src/result, re-export from logger
tapErr is a Result combinator — it takes any `(err: E) => void`, not a
logger method. Living in src/logger/ was motivated by the use case,
not the dependencies. Moving it next to isErr/tryAsync where it
belongs, and re-exporting from wellcrafted/logger so the canonical
`tapErr(log.warn)` idiom still imports cleanly.
