+++ src/result/index.ts
@@ -1,2 +1,3 @@
 export * from "./result.js";
+export * from "./tap-err.js";
 export * from "./utils.js";
